requirements.txt
streamlit
python-dotenv
PyPDF2
pytesseract
pdf2image
langchain
langchain-openai
langchain-community
openai
